+++ docs/getting-started/changelog.md
@@ -6,6 +6,10 @@ Components with the <lynk-badge type="warning" pill>Experimental</lynk-badge> ba
 
 <lynk-alert type="info" open>During the beta period, these restrictions may be relaxed in the event of a mission-critical bug. 🐛</lynk-alert>
 
+## 0.9.12
+- Adjusting styles of `<lynk-button>` outline styles and making disabled slightly more obvious.
+- 🎉 NEW: Added a `<lynk-dot-loader>` component for another loader style.
+
 ## 0.9.11
 - Added some additional style options to  `<lynk-tag>`.